Click on widget title opens the board in Trello
Like the card change earlier, we can take advantage of the Trello app's
URL intercepting feature that will grab https://trello.com/* and redirect
to the appropriate board view for us.

If Trello is not installed, it opens the board in a browser.

Note that this is only a partial solution - As it stands we only have
access to the board object itself during widget configuration, so any
existing widgets will have to be removed and re-added for their title
intent to be changed appropriately.  If they are not, the default behavior
remains: opening the top-level trello app or the main trello.com site.

Signed-off-by: Jim Ramsay <i.am@jimramsay.com>