Fix ColorPreference SVBar when loading black colors
Turns out there's a defect in com.larswerkman.holocolorpicker.ColorPicker
when dealing with black colors, and this works around it until the
underlying issue is fixed in that library.

With this change the SVBar actually jumps to the right location when the
color picker dialog is opened even for black when previously it would jump
to the center (fully saturated color) if the actual color was black.

Signed-off-by: Jim Ramsay <i.am@jimramsay.com>