Add viewModel for ConfigActivity
Added a viewModel to the activity to make it a proper UI controller. All UI logic is then handled by the viewModel.
Use LiveData to observe the response through the UI controller.
Set the listener of the activity's buttons explicitly in the activity.
Added a DataStatus class as a wrapper for the data fetched from the remote, then the TrelloApi handles all the calls and JSON parsing and just returns the data the UI needs.
Extracted Volley stuff into a new class.